+++ packages/core-react/src/index.test.ts
@@ -1,6 +1,6 @@
 import { expect, it } from "vitest";
 
-import * as Exports from ".";
+import * as Exports from "./";
 
 it("should expose correct exports", () => {
   expect(Object.keys(Exports).sort()).toMatchInlineSnapshot(`
