--- packages/core-web/src/crypto.ts
@@ -1,5 +1,0 @@
-export {
-  importPKCS8,
-  signAccessJwt,
-  type SignAccessJwtOptions,
-} from "@livepeer/core/crypto";
