+++ packages/core-web/src/media.test.ts
@@ -5,13 +5,8 @@ import * as Exports from "./media";
 it("should expose correct exports", () => {
   expect(Object.keys(Exports).sort()).toMatchInlineSnapshot(`
     [
-      "DEFAULT_AUTOHIDE_TIME",
-      "DEFAULT_VOLUME_LEVEL",
-      "addMediaMetricsToStore",
-      "aspectRatios",
       "createControllerStore",
       "getMediaSourceType",
-      "sanitizeMediaControllerState",
     ]
   `);
 });
