+++ packages/core-web/src/media/metrics.test.ts
@@ -1,11 +1,10 @@
 import { beforeAll, describe, expect, it } from "vitest";
 
-import { MockedVideoElement, resetDateNow, setupClient } from "../../test";
+import { MockedVideoElement, resetDateNow } from "../../test";
 import { addMediaMetrics } from "./metrics";
 
 describe("addMediaMetrics", () => {
   beforeAll(() => {
-    setupClient();
     resetDateNow();
   });
 
@@ -25,32 +24,32 @@ describe("addMediaMetrics", () => {
       }
 
       expect(metricsSnapshot?.current).toMatchInlineSnapshot(`
-          {
-            "autoplay": "standard",
-            "duration": null,
-            "firstPlayback": null,
-            "nError": null,
-            "nStalled": 0,
-            "nWaiting": 0,
-            "offset": null,
-            "pageUrl": "http://localhost:3000/",
-            "playbackScore": null,
-            "player": "hls-1",
-            "playerHeight": null,
-            "playerWidth": null,
-            "preloadTime": null,
-            "sourceType": "unknown",
-            "sourceUrl": null,
-            "timeStalled": 0,
-            "timeUnpaused": 0,
-            "timeWaiting": 0,
-            "ttff": null,
-            "uid": "",
-            "userAgent": "UA",
-            "videoHeight": null,
-            "videoWidth": null,
-          }
-        `);
+        {
+          "autoplay": "preload-full",
+          "duration": null,
+          "firstPlayback": null,
+          "nError": null,
+          "nStalled": 0,
+          "nWaiting": 0,
+          "offset": null,
+          "pageUrl": "http://localhost:3000/",
+          "playbackScore": null,
+          "player": "hls-1",
+          "playerHeight": null,
+          "playerWidth": null,
+          "preloadTime": null,
+          "sourceType": "unknown",
+          "sourceUrl": null,
+          "timeStalled": 0,
+          "timeUnpaused": 0,
+          "timeWaiting": 0,
+          "ttff": null,
+          "uid": "",
+          "userAgent": "UA",
+          "videoHeight": null,
+          "videoWidth": null,
+        }
+      `);
     });
 
     it("should update time unpaused and first playback", async () => {
@@ -71,7 +70,7 @@ describe("addMediaMetrics", () => {
 
       expect(metricsSnapshot?.current).toMatchInlineSnapshot(`
         {
-          "autoplay": "standard",
+          "autoplay": "preload-full",
           "duration": null,
           "firstPlayback": null,
           "nError": null,
@@ -115,32 +114,32 @@ describe("addMediaMetrics", () => {
       }
 
       expect(metricsSnapshot?.current).toMatchInlineSnapshot(`
-      {
-        "autoplay": "standard",
-        "duration": null,
-        "firstPlayback": null,
-        "nError": null,
-        "nStalled": 0,
-        "nWaiting": 1,
-        "offset": null,
-        "pageUrl": "http://localhost:3000/",
-        "playbackScore": null,
-        "player": "hls-1",
-        "playerHeight": null,
-        "playerWidth": null,
-        "preloadTime": null,
-        "sourceType": "unknown",
-        "sourceUrl": null,
-        "timeStalled": 0,
-        "timeUnpaused": 0,
-        "timeWaiting": 1000,
-        "ttff": null,
-        "uid": "",
-        "userAgent": "UA",
-        "videoHeight": null,
-        "videoWidth": null,
-      }
-    `);
+        {
+          "autoplay": "preload-full",
+          "duration": null,
+          "firstPlayback": null,
+          "nError": null,
+          "nStalled": 0,
+          "nWaiting": 1,
+          "offset": null,
+          "pageUrl": "http://localhost:3000/",
+          "playbackScore": null,
+          "player": "hls-1",
+          "playerHeight": null,
+          "playerWidth": null,
+          "preloadTime": null,
+          "sourceType": "unknown",
+          "sourceUrl": null,
+          "timeStalled": 0,
+          "timeUnpaused": 0,
+          "timeWaiting": 1000,
+          "ttff": null,
+          "uid": "",
+          "userAgent": "UA",
+          "videoHeight": null,
+          "videoWidth": null,
+        }
+      `);
     });
 
     it("should update time stalled and stalled count", async () => {
@@ -162,32 +161,32 @@ describe("addMediaMetrics", () => {
       }
 
       expect(metricsSnapshot?.current).toMatchInlineSnapshot(`
-                    {
-                      "autoplay": "standard",
-                      "duration": null,
-                      "firstPlayback": null,
-                      "nError": null,
-                      "nStalled": 1,
-                      "nWaiting": 0,
-                      "offset": null,
-                      "pageUrl": "http://localhost:3000/",
-                      "playbackScore": null,
-                      "player": "hls-1",
-                      "playerHeight": null,
-                      "playerWidth": null,
-                      "preloadTime": null,
-                      "sourceType": "unknown",
-                      "sourceUrl": null,
-                      "timeStalled": 1000,
-                      "timeUnpaused": 0,
-                      "timeWaiting": 0,
-                      "ttff": null,
-                      "uid": "",
-                      "userAgent": "UA",
-                      "videoHeight": null,
-                      "videoWidth": null,
-                    }
-                  `);
+        {
+          "autoplay": "preload-full",
+          "duration": null,
+          "firstPlayback": null,
+          "nError": null,
+          "nStalled": 1,
+          "nWaiting": 0,
+          "offset": null,
+          "pageUrl": "http://localhost:3000/",
+          "playbackScore": null,
+          "player": "hls-1",
+          "playerHeight": null,
+          "playerWidth": null,
+          "preloadTime": null,
+          "sourceType": "unknown",
+          "sourceUrl": null,
+          "timeStalled": 1000,
+          "timeUnpaused": 0,
+          "timeWaiting": 0,
+          "ttff": null,
+          "uid": "",
+          "userAgent": "UA",
+          "videoHeight": null,
+          "videoWidth": null,
+        }
+      `);
     });
   });
 
@@ -212,31 +211,31 @@ describe("addMediaMetrics", () => {
     }
 
     expect(metricsSnapshot?.current).toMatchInlineSnapshot(`
-                    {
-                      "autoplay": "standard",
-                      "duration": null,
-                      "firstPlayback": null,
-                      "nError": null,
-                      "nStalled": 0,
-                      "nWaiting": 0,
-                      "offset": null,
-                      "pageUrl": "http://localhost:3000/",
-                      "playbackScore": null,
-                      "player": "hls-1",
-                      "playerHeight": null,
-                      "playerWidth": null,
-                      "preloadTime": null,
-                      "sourceType": "unknown",
-                      "sourceUrl": null,
-                      "timeStalled": 0,
-                      "timeUnpaused": 0,
-                      "timeWaiting": 0,
-                      "ttff": null,
-                      "uid": "",
-                      "userAgent": "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.101 Safari/537.36",
-                      "videoHeight": null,
-                      "videoWidth": null,
-                    }
-                  `);
+      {
+        "autoplay": "preload-full",
+        "duration": null,
+        "firstPlayback": null,
+        "nError": null,
+        "nStalled": 0,
+        "nWaiting": 0,
+        "offset": null,
+        "pageUrl": "http://localhost:3000/",
+        "playbackScore": null,
+        "player": "hls-1",
+        "playerHeight": null,
+        "playerWidth": null,
+        "preloadTime": null,
+        "sourceType": "unknown",
+        "sourceUrl": null,
+        "timeStalled": 0,
+        "timeUnpaused": 0,
+        "timeWaiting": 0,
+        "ttff": null,
+        "uid": "",
+        "userAgent": "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.101 Safari/537.36",
+        "videoHeight": null,
+        "videoWidth": null,
+      }
+    `);
   });
 });
