--- packages/core-web/src/types.ts
@@ -1,1 +1,0 @@
-export type { Address, Hash } from "@livepeer/core/types";
