--- packages/core-web/src/utils.ts
@@ -1,15 +1,0 @@
-export * from "@livepeer/core/utils";
-
-export {
-  b64Decode,
-  b64Encode,
-  b64UrlDecode,
-  b64UrlEncode,
-  deepMerge,
-  noop,
-  omit,
-  parseArweaveTxId,
-  parseCid,
-  pick,
-  warn,
-} from "@livepeer/core/utils";
