+++ packages/core-web/src/webrtc.test.ts
@@ -5,13 +5,12 @@ import * as Exports from "./webrtc";
 it("should expose correct exports", () => {
   expect(Object.keys(Exports).sort()).toMatchInlineSnapshot(`
     [
-      "changeMediaStream",
+      "attachMediaStreamToPeerConnection",
       "createNewWHEP",
       "createNewWHIP",
       "getDisplayMedia",
       "getMediaDevices",
       "getUserMedia",
-      "isWebRTCSupported",
     ]
   `);
 });
