+++ packages/core-web/src/webrtc/shared.ts
@@ -1,7 +1,5 @@
-import {
-  AccessControlParams,
-  NOT_ACCEPTABLE_ERROR_MESSAGE,
-} from "@livepeer/core";
+import { NOT_ACCEPTABLE_ERROR_MESSAGE } from "@livepeer/core/errors";
+import { AccessControlParams } from "@livepeer/core/media";
 import { isClient } from "../media/utils";
 
 /**
