+++ packages/core/src/errors.test.ts
@@ -1,6 +1,6 @@
 import { expect, it } from "vitest";
 
-import * as Exports from "./";
+import * as Exports from "./errors";
 
 it("should expose correct exports", () => {
   expect(Object.keys(Exports).sort()).toMatchInlineSnapshot(`
@@ -8,25 +8,14 @@ it("should expose correct exports", () => {
       "ACCESS_CONTROL_ERROR_MESSAGE",
       "BFRAMES_ERROR_MESSAGE",
       "NOT_ACCEPTABLE_ERROR_MESSAGE",
+      "PERMISSIONS_ERROR_MESSAGE",
       "STREAM_OFFLINE_ERROR_MESSAGE",
       "STREAM_OPEN_ERROR_MESSAGE",
-      "b64Decode",
-      "b64Encode",
-      "b64UrlDecode",
-      "b64UrlEncode",
-      "createControllerStore",
-      "createStorage",
-      "deepMerge",
-      "getMediaSourceType",
       "isAccessControlError",
       "isBframesError",
       "isNotAcceptableError",
+      "isPermissionsError",
       "isStreamOfflineError",
-      "noopStorage",
-      "omit",
-      "pick",
-      "sanitizeMediaControllerState",
-      "version",
     ]
   `);
 });
