+++ packages/core/src/external.test.ts
@@ -1,15 +1,11 @@
 import { expect, it } from "vitest";
 
-import * as Exports from "./crypto";
-
-globalThis.crypto = crypto as Crypto;
-globalThis.window.crypto = crypto as Crypto;
+import * as Exports from "./external";
 
 it("should expose correct exports", () => {
   expect(Object.keys(Exports).sort()).toMatchInlineSnapshot(`
     [
-      "importPKCS8",
-      "signAccessJwt",
+      "getSrc",
     ]
   `);
 });
