--- packages/core/src/index.ts
@@ -1,65 +1,0 @@
-export {
-  ACCESS_CONTROL_ERROR_MESSAGE,
-  BFRAMES_ERROR_MESSAGE,
-  NOT_ACCEPTABLE_ERROR_MESSAGE,
-  PERMISSIONS_ERROR_MESSAGE,
-  STREAM_OFFLINE_ERROR_MESSAGE,
-  STREAM_OPEN_ERROR_MESSAGE,
-  isAccessControlError,
-  isBframesError,
-  isNotAcceptableError,
-  isPermissionsError,
-  isStreamOfflineError,
-} from "./errors";
-export {
-  createControllerStore,
-  getMediaSourceType,
-  sanitizeMediaControllerState,
-} from "./media";
-export type {
-  AccessControlParams,
-  AriaText,
-  AudioSrc,
-  AudioTrackSelector,
-  Base64Src,
-  ClipLength,
-  ClipParams,
-  ControlsOptions,
-  ControlsState,
-  DeviceInformation,
-  ElementSize,
-  HlsSrc,
-  InitialProps,
-  MediaControllerCallbackState,
-  MediaControllerState,
-  MediaControllerStore,
-  MediaMetrics,
-  MediaSizing,
-  Metadata,
-  MetricsStatus,
-  ObjectFit,
-  PlaybackError,
-  PlaybackMonitor,
-  PlaybackRate,
-  SingleAudioTrackSelector,
-  SingleTrackSelector,
-  SingleVideoTrackSelector,
-  Src,
-  VideoQuality,
-  VideoSrc,
-  VideoTrackSelector,
-  WebRTCSrc,
-} from "./media";
-export { createStorage, noopStorage } from "./storage";
-export type { ClientStorage as Storage } from "./storage";
-export type { Address, Hash } from "./types";
-export {
-  b64Decode,
-  b64Encode,
-  b64UrlDecode,
-  b64UrlEncode,
-  deepMerge,
-  omit,
-  pick,
-} from "./utils";
-export { version } from "./version";
