+++ packages/react/src/context.test.ts
@@ -1,15 +1,14 @@
 import { expect, it } from "vitest";
 
-import * as Exports from "./crypto";
-
-globalThis.crypto = crypto as Crypto;
-globalThis.window.crypto = crypto as Crypto;
+import * as Exports from "./context";
 
 it("should expose correct exports", () => {
   expect(Object.keys(Exports).sort()).toMatchInlineSnapshot(`
     [
-      "importPKCS8",
-      "signAccessJwt",
+      "MediaProvider",
+      "createMediaScope",
+      "useMediaContext",
+      "useStore",
     ]
   `);
 });
