+++ packages/react/src/crypto.ts
@@ -2,4 +2,4 @@ export {
   importPKCS8,
   signAccessJwt,
   type SignAccessJwtOptions,
-} from "@livepeer/core-react/crypto";
+} from "@livepeer/core/crypto";
