+++ vitest.config.ts
@@ -1,9 +1,16 @@
-import { defineConfig } from "vitest/config";
+import { configDefaults, defineConfig } from "vitest/config";
 
 export default defineConfig({
   test: {
     coverage: {
       reporter: ["text", "json", "html"],
+      exclude: [
+        ...configDefaults.exclude,
+        "**/examples/**",
+        "**/test/**",
+        "**/*.d.ts",
+        "**/generate-version.ts",
+      ],
     },
     environment: "jsdom",
     setupFiles: [
